requirement.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langchain-openai
qdrant-client
pypdf
python-dotenv
httpx
langchain-qdrant
